added routes/movie.js
